Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int n;
+    float n;
     scanf("",&n);
-    printf("%d",n);
+    printf("%f",n);
     return 0;
 }
